refactor(member): memberType 이동 (member -> global) - 인증, 인가 관련 패키지(Auth)에 사용하기 위해 global로 이동
